DSFAAP-2151: separate out existing API from alpha API types
