+++ src/types/holdings.js
@@ -1,8 +1,8 @@
 import Joi from 'joi'
 
 import { LinksReference } from './links.js'
-// import { LocationsReference } from './locations.js'
-// import { CustomersReference } from './customers.js'
+import { LocationsReference } from './locations.js'
+import { CustomersReference } from './customers.js'
 
 const HoldingsData = Joi.object({
   type: Joi.string()
@@ -20,11 +20,8 @@ export const HoldingsReference = Joi.object({
 
 export const Holdings = HoldingsData.keys({
   cphType: Joi.string().required().label('CPH Type'),
-  localAuthority: Joi.string().label('Local Authority'),
   relationships: Joi.object({
-    // DSFAAP-2110 - not needed
-    // > removed, because they aren't strictly necessary for the first WFM use case.
-    // location: LocationsReference.optional(),
-    // cphHolder: CustomersReference.optional()
+    location: LocationsReference.optional(),
+    cphHolder: CustomersReference.optional()
   })
 }).label('Holdings Data')
