+++ src/types/locations.js
@@ -1,6 +1,7 @@
 import Joi from 'joi'
 
-import { Commodities, CommoditiesData } from './commodities.js'
+import { CommoditiesData } from './commodities.js'
+import { FacilitiesData } from './facilities.js'
 import { LinksReference } from './links.js'
 
 export const LocationsData = Joi.object({
@@ -17,7 +18,7 @@ export const LocationsReference = Joi.object({
   links: LinksReference
 })
 
-export const Address = Joi.object({
+const Address = Joi.object({
   paonStartNumber: Joi.alternatives(Joi.number(), Joi.string()).allow(null, ''),
   paonStartNumberSuffix: Joi.string().allow(null, ''),
   paonEndNumber: Joi.alternatives(Joi.number(), Joi.string()).allow(null, ''),
@@ -31,11 +32,7 @@ export const Address = Joi.object({
   street: Joi.string().allow(null, ''),
   locality: Joi.string().allow(null, ''),
   town: Joi.string().allow(null, ''),
-
-  // DSFAAP-2110 - not needed
-  // > commented because not strictly necessary for WFM first use case
-  // > unless this is what is meant by local authority
-  // administrativeAreaCounty: Joi.string().allow(null, ''), // maps ADMINISTRATIVE_AREA
+  administrativeAreaCounty: Joi.string().allow(null, ''), // maps ADMINISTRATIVE_AREA
   postcode: Joi.string().allow(null, ''),
   ukInternalCode: Joi.string().allow(null, ''),
   countryCode: Joi.string().allow(null, '')
@@ -43,30 +40,14 @@ export const Address = Joi.object({
 
 export const Locations = LocationsData.keys({
   address: Address.required(),
-  osMapReference: Joi.string().allow(null, '').label('OS map reference'),
   relationships: Joi.object({
     commodities: Joi.object({
-      data: Joi.alternatives(
-        CommoditiesData,
-        Joi.array().items(CommoditiesData)
-      ).required(),
+      data: Joi.array().items(CommoditiesData).required(),
+      links: LinksReference
+    }),
+    facilities: Joi.object({
+      data: Joi.array().items(FacilitiesData).required(),
       links: LinksReference
     })
-    // DSFAAP-2110 - not needed
-    // > only the id is needed, which can be acquired from the workorder directly
-    // facilities: Joi.object({
-    //   data: Joi.alternatives(
-    //     FacilitiesData,
-    //     Joi.array().items(FacilitiesData)
-    //   ).required(),
-    //   links: LinksReference
-    // })
   })
-})
-
-export const LocationsHydrated = Locations.keys({
-  livestockUnits: Joi.array().items(Commodities).required()
-  // DSFAAP-2110 - not needed
-  // > only the id is needed, which can be acquired from the workorder directly
-  // facilities: Joi.array().items(FacilitiesData).required()
 })
